Extract batchSpecWorkspaceCreator into file
